Add timestamping
I redid the proto file, changing User to Process, as both the server and the clients needed to be represented with id's and local Lamport Clocks which they get with a timestamp field. LogMessage is removed as it is not going to be needed for logging events which only the server should do, whereas the client just logs messages.

Connect rpc now takes a UserName, which is just to set the name of the client as it is presented to others

In the server we checked for validity of messages. This functionality has been moved to the Messages file and redone to work on strings instead, as we don't want to create a chatMessage for invalid strings

field accessors of grpc generated elements has been set to use getters, which I suspect is safer/cleaner.

Disconnect events prompts the server to call the rpc method of SendChat, because it is a message to be sent to all participants

timestamping works via lamport clocks facilitated by functions in the ProcessExtension file, where they will manipulate the timestamp according to the lamport clock pseudocode as presented in slides.

For now the timestamps lock when the functions for incrementing and setting the timestamps of the processes, such that we avoid race conditions